requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
Flask==2.2.3
fonttools==4.39.3
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0+cpu
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.3
